Add auto_date and revision_hash options
This allows users, to use the preset snippets $DATE
and $REVISION in their projects. This is great for title pages.

For both the new options, I just picked an arbitrary $STRING
and then used the process_snippets function to search for them,
 if the users had enabled the correct setting.

To implement auto_date we I thought we needed a date library,
so I picked dayjs because its small and will allow us to, at some point,
 add customsiable date strings.

To implement revision_hash I just used a child process to call git
and print the most recent hash to stdout.

See-Also:
https://stackoverflow.com/a/35778030/1649917
https://github.com/you-dont-need/You-Dont-Need-Momentjs
https://github.com/iamkun/dayjs